curiosity-intelligence/requirements.txt
# Curiosity Intelligence Engine - Dependencies

# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
asyncio>=3.4.3

# OpenAI for embeddings
openai>=1.0.0

# Database
supabase>=2.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pgvector>=0.2.0

# Reddit API
praw>=7.7.0

# Clustering
hdbscan>=0.8.33
numpy>=1.24.0
scikit-learn>=1.3.0

# Text processing
regex>=2023.0.0

# Email rendering
mjml>=0.5.0
jinja2>=3.1.0

# Image generation (share cards)
pillow>=10.0.0

# CLI
click>=8.1.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0
resend>=2.0.0
